Breaking change: regex handling
Previously, we'd split the messages into words and feed each word to the
user-configured free filters.
This turned out to not be ideal, as sometimes we want more control with
the regex, including capitalisation.

With this change, any free filter that starts with '/' will be expected
to handle the full message string.